refactor: extract shared ChatTextarea component
The hero chat input and AI chat input bar duplicated textarea handling
logic (text state, IME-aware keydown, form submit, auto-grow, icon
button styles). Extract a shared ChatTextarea component to consolidate
this — especially the IME composition handling — so it only needs to
be maintained in one place.
